% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectiveMLE.R
\name{compute_likelihood_function}
\alias{compute_likelihood_function}
\title{Compute the Likelihood Function for Beta1 Optimization}
\usage{
compute_likelihood_function(
  X,
  y,
  sigma_sq,
  alpha_ov = 0.05,
  B = 1e+06,
  seed = 12345
)
}
\arguments{
\item{X}{A matrix of predictor variables.}

\item{y}{A vector of response variables.}

\item{sigma_sq}{The variance parameter.}

\item{alpha_ov}{The significance level of the overall F-test.}

\item{B}{The number of Monte Carlo samples (default is 100000).}

\item{seed}{A seed value for reproducibility (default is 12345).}
}
\value{
A function that computes the (log) likelihood for a given `beta1`.
}
\description{
This function constructs a likelihood function (to be optimized) using the given data.
The likelihood is computed via a Monte Carlo integration method that involves generating
random samples from chi-squared distributions.
}
