% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{get_Xy_centered}
\alias{get_Xy_centered}
\title{Center X and y by Removing the Intercept}
\usage{
get_Xy_centered(X, y)
}
\arguments{
\item{X}{A numeric matrix of predictors (dimensions: n x p).}

\item{y}{A numeric response vector of length n.}
}
\value{
A list containing:
\describe{
  \item{X}{The transformed predictor matrix with the intercept projected out.}
  \item{y}{The transformed response vector with the intercept projected out.}
}
}
\description{
This function centers the predictor matrix `X` and response vector `y` by
projecting out the intercept. It ensures that the transformed data does not
include a mean shift component.
}
\details{
This function uses singular value decomposition (SVD) to project out
the intercept and return a centered version of `X` and `y`.
}
\examples{
X <- matrix(rnorm(100), nrow = 10)
y <- rnorm(10)
centered_data <- get_Xy_centered(X, y)
X_centered <- centered_data$X
y_centered <- centered_data$y

}
