% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmFScreen.R
\name{lmFScreen.fit}
\alias{lmFScreen.fit}
\title{This function takes as input a design matrix X and output vector Y and fits a linear regression model. It then conducts F-screening by
(1) testing the overall hypothesis that all coefficients in the linear regression are zero using an F-test, and
(2) if this overall test is rejected, it outputs selective p-values, confidence intervals, and point estimates for the coefficients in the linear regression model.
If the overall test is not rejected, it returns the overall F-statistic and indicates that it is not significant.}
\usage{
lmFScreen.fit(
  X,
  y,
  alpha = 0.05,
  alpha_ov = 0.05,
  test_cols = 1:ncol(X),
  sigma_sq = NULL,
  compute_CI = TRUE,
  compute_est = TRUE,
  B = 1e+05
)
}
\arguments{
\item{X}{A numeric matrix of predictors.}

\item{y}{A numeric response vector.}

\item{alpha}{Significance level for confidence intervals and hypothesis tests (default: 0.05).}

\item{alpha_ov}{Significance level for the overall F-test used for screening (default: 0.05).}

\item{test_cols}{Indices of predictors to test (default: all columns of X).}

\item{sigma_sq}{Optional noise variance. If NULL, it is estimated using a corrected residual variance.}

\item{compute_CI}{Logical; whether to compute selective confidence intervals (default: TRUE).}

\item{compute_est}{Logical; whether to compute selective point estimates (default: TRUE).}

\item{B}{Number of Monte Carlo samples used for selective inference (default: 100000).}
}
\value{
A list of class lmFScreen containing:
\itemize{
\item Selective coefficients, confidence intervals, and p-values
\item Standard (OLS) coefficients, confidence intervals, and p-values
\item Model settings: alpha and alpha_ov
}
}
\description{
This function takes as input a design matrix X and output vector Y and fits a linear regression model. It then conducts F-screening by
(1) testing the overall hypothesis that all coefficients in the linear regression are zero using an F-test, and
(2) if this overall test is rejected, it outputs selective p-values, confidence intervals, and point estimates for the coefficients in the linear regression model.
If the overall test is not rejected, it returns the overall F-statistic and indicates that it is not significant.
}
\examples{
data(mtcars)
X <- cbind(mtcars$wt, mtcars$hp)
y <- mtcars$mpg
svdP <- svd(rep(1,nrow(mtcars)), nu = nrow(mtcars))
tol <- nrow(mtcars) * max(svdP$d) * .Machine$double.eps
r <- sum(svdP$d > tol)
U_full <- svdP$u
U_perp <- U_full[, (r+1):ncol(U_full)]
X <- t(U_perp) \%*\% X
y <- t(U_perp) \%*\% y
result <- lmFScreen.fit(X,y)
summary(result)
coef(result)
confint(result)

}
