% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{confint.lmFScreen}
\alias{confint.lmFScreen}
\title{Compute Confidence Intervals for an `lmFScreen` Model}
\usage{
\method{confint}{lmFScreen}(object, ...)
}
\arguments{
\item{object}{An object of class `"lmFScreen"`, containing model results.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns a list containing:
\describe{
  \item{Selective}{A dataframe with confidence intervals for selective estimates.}
  \item{Naive}{A dataframe with confidence intervals for naive estimates.}
}
}
\description{
This function returns and prints confidence intervals for both **selective** and **naive** estimates.
}
