% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{get_Xy_centered}
\alias{get_Xy_centered}
\title{Project Out the Intercept from X and y}
\usage{
get_Xy_centered(X, y)
}
\arguments{
\item{X}{A numeric matrix of predictors with dimensions (n x p).}

\item{y}{A numeric response vector of length n.}
}
\value{
A list containing:
\describe{
\item{X}{The predictor matrix after projection (dimensions: (n - 1) x p).}
\item{y}{The response vector after projection (length: n - 1).}
}
}
\description{
This function projects the predictor matrix \code{X} and response vector \code{y} onto the
orthogonal complement of the intercept (constant) vector. This removes any component
of the data explained by the intercept, effectively centering the data in the
sense of removing the mean direction.
}
\details{
This function uses the singular value decomposition (SVD) of the intercept vector
to construct an orthonormal basis for the space orthogonal to the intercept.
It then projects \code{X} and \code{y} onto that space, effectively removing any contribution
from the intercept.
}
\examples{
X <- matrix(rnorm(100), nrow = 10)
y <- rnorm(10)
centered_data <- lmFScreen:::get_Xy_centered(X, y)
X_centered <- centered_data$X
y_centered <- centered_data$y

}
